mm5btc_bot/requirements.txt
aiogram==3.15.0
python-dotenv==1.0.1
watchfiles==1.0.4
qrcode==8.2
Pillow==12.1.1
